+++ apps/backend/src/libs/types/types.ts
@@ -2,7 +2,6 @@ export { type Entity } from "./entity.type.js";
 export { type Repository } from "./repository.type.js";
 export { type Service } from "./service.type.js";
 export {
-	EMPTY_ARRAY_LENGTH,
 	type PaginationRequestDto,
 	type PaginationResponseDto,
 	type ServerCommonErrorResponse,