+++ apps/frontend/src/libs/components/course/course.tsx
@@ -10,6 +10,7 @@ import {
 import { Button } from "../button/button.js";
 import { Link } from "../link/link.js";
 import { CourseCard } from "./libs/component/component.js";
+import { checkIsUserCourse } from "./libs/helpers/helpers.js";
 import styles from "./styles.module.css";
 
 type Properties = {
@@ -37,8 +38,17 @@ const Course: React.FC<Properties> = ({
 			userId: String(userId),
 		},
 	) as typeof AppRoute.USERS_$USER_ID_COURSES_$COURSE_ID;
+	const courseComparisonRoute = configureString(
+		AppRoute.USERS_$USER_ID_COURSES_$COURSE_ID_COMPARE,
+		{
+			courseId: String(id),
+			userId: String(userId),
+		},
+	) as typeof AppRoute.USERS_$USER_ID_COURSES_$COURSE_ID_COMPARE;
 
 	const hasAddCourse = !!onAddCourse;
+	const hasProgress = checkIsUserCourse(course);
+	const hasCompare = !hasProgress && !hasAddCourse;
 
 	const handleAddCourse = useCallback(() => {
 		onAddCourse?.({
@@ -49,12 +59,12 @@ const Course: React.FC<Properties> = ({
 
 	return (
 		<article className={styles["container"]}>
-			{hasAddCourse ? (
-				<CourseCard course={course} />
-			) : (
+			{hasProgress ? (
 				<Link className={styles["link"]} to={courseDescriptionRouteById}>
 					<CourseCard course={course} />
 				</Link>
+			) : (
+				<CourseCard course={course} />
 			)}
 			{hasAddCourse && (
 				<div className={styles["actions"]}>
@@ -72,6 +82,16 @@ const Course: React.FC<Properties> = ({
 						isDisabled={isButtonDisabled}
 						label="Add"
 						onClick={handleAddCourse}
+						size="small"
+					/>
+				</div>
+			)}
+			{hasCompare && (
+				<div className={styles["actions"]}>
+					<Button
+						className={styles["compare-progress-button"]}
+						href={courseComparisonRoute}
+						label="Compare progress"
 						size="small"
 					/>
 				</div>