+++ packages/shared/src/modules/user-courses/libs/enums/user-courses-api-path.enum.ts
@@ -1,5 +1,6 @@
 const UserCoursesApiPath = {
 	$USER_ID: "/:userId",
+	$USER_ID_COMMON: "/:userId/common",
 	ROOT: "/",
 } as const;